feat: fix prices

+++ internal/allocation/models.go
@@ -11,8 +11,8 @@ import (
 type TStrategyAllocation struct {
 	Strategy      *models.TStrategyList              `json:"strategy"`
 	RiskGroup     *strategies.TStrategyGroupFromRisk `json:"riskGroup"`
-	CurrentTVL    *big.Int                           `json:"currentTVL"` // value in WANT
-	AvailableTVL  *big.Int                           `json:"availableTVL"`
-	CurrentUSDC   *big.Int                           `json:"currencyUSDC"` // value in USDC
-	AvailableUSDC *big.Int                           `json:"availableUSDC"`
+	CurrentTVL    *big.Float                         `json:"currentTVL"` // value in WANT
+	AvailableTVL  *big.Float                         `json:"availableTVL"`
+	CurrentUSDC   *big.Float                         `json:"currencyUSDC"` // value in USDC
+	AvailableUSDC *big.Float                         `json:"availableUSDC"`
 }
